pyproject.toml
[project]
name = "Nonebot_AyaSanKo"
version = "1.0.0"
description = "Nonebot_AyaSanKo"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[httpx,websockets]>=2.4.4",
    "nonebot-adapter-qq>=1.6.6",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
]

[tool.nonebot]
plugin_dirs = []
builtin_plugins = ["echo", "single_session"]

[tool.nonebot.adapters]
nonebot-adapter-qq = [{ name = "QQ", module_name = "nonebot.adapters.qq" }]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
